Add minimal framework for requesting Polymarket for Sampling Markets.
This commit adds the basic function of requesting Polymarket for all
tradeable markets using the following CLOB API guide:
https://docs.polymarket.com/#get-simplified-markets.

A new type [Market.t] is linted, to provide a structured way to
interpret the data. Furthermore, a helper type [Next_cursor.t] is
created to deal with the pagination of requesting their API, as only 500
markets at a time is given. Then, a pointer in the form of a
[Next_cursor.t], akin to a linked list, would give you the next URL to
request for the next 500 markets.

Future commits will populate the [Market.t] type.

JSON parsing is not tested as it is not quite within our control. It
highly depends on the structure of the responses given by Polymarket,
which must be trusted.
Furthermore, it uses the Yojson library, which we can safely assume is
heavily tested and all functions work exactly as advertised.