Lint Token and Price types, rehaul of_json signature
This commit lints the Token and Price type, which are required to
populate the Market type, based on the API as follows:
https://docs.polymarket.com/#markets
It should be noted that the Token type does not return the price and the
winner according to the API documentation, but when requested, do return
the price and winner.
We take it for granted for now, but it could be considered incorrect
behaviour if the API does retract that one day.

The Price type is linted to deal with monetary values, as storing
currency as a float rarely goes well. We use a subdivision of $0.0001
cents, as this seems to be the division Polymarket uses under the hood,
based on the API queries.

The [of_json] functions have also been slightly adjusted, to accept a
key string if the data type is easily inferrable using the in-built
functions in [Yojson.Basic.Util]. Otherwise, it takes a tuple,
requesting the user to provide a function to interpret the JSON as the
custom type. It should be noted that in [Price] and [Next_cursor], the
[of_json] functions do not do any validation that the JSON indeed
contains the float or string as requested, as we must trust the
developer to not misuse the function.