Add price of token endpoint and lint a new Endpoint type
This commit exposes the get price endpoint as seen from the
documentation as follows: https://docs.polymarket.com/#get-price.

This returns a price given a [token_id] and a [Side], which we lint a module
for since it is clearly a variant type with [Buy] or [Sell].

By adding another endpoint, some portions of code duplication is made
obvious and refactoring of logic needs to be done. In this case.
[endpoint.ml] is a new file that helps us factor out the CLOB URI as
well as the querying for JSON.

Simple testing is done to ensure that the URI queried is correct.