Populate [Market] type with useful fields
I anticipate that the added fields of [closed], [condition_id] (market
id), [description], [end_date] and [icon] can be useful in the future
and are useful things to parse. Hence, this commit adds them to the
record.

A noteworthy point is that the [end_date] could be [null], so this is
represented with an [option].