Lint Portfolio type and rename Price to Dollar
This commit creates a portfolio type representing the assets a person
holds. A portfolio contains an amount of money and the number of token IDs they
own.

As a result, the Price module has been renamed to Dollar to generalise
the use of the module without changing much code.

[Token.Id.t] is also a new type linted to provide more compile time
guarantees about the string representing the token.