Lint Ledger type
This commit adds the [Ledger.t] type. This provides a way to interact
with an abstraction over a mapping between [Person.t]s and their
[Portfolio.t]s.

A future commit can then be added which directly saves this [Ledger.t]
type to a file for persistent memory.

The [ledger] folder is renamed to [asset_management] for a more general
meaning, freeing up the name to be used by [ledger.ml].

The [Transaction_failure.t] has also been refactored out into its own
file as both the [Portfolio] and [Ledger] can raise their own errors.

The [Portfolio.remove] function is also now capped at zero instead of
letting it go negative. This is more in line with expectations of how
the function should work.