Populate dune exec command with API interactions
This commit adds Polymarket querying to the dune exec command. Names of
modules are changed to better explain what they do. For example, instead
of [client], [cli] is used, and the commands are given appropriate names
of their function.

[arg_type] is exposed in the various types that are used as command line
parameters.

Notably, this commit fixes [get_price.ml], as it was previously
untested. The price returned by the JSON in [get_price] is a string
while the price returned by the JSON in [sampling_markets] is a float.
Hence, in the [dollar.ml] module, I have exposed both an [of_json_float]
and [of_json_string] function to help with this issue.