Add commands to interact with persistent ledger
This commit adds a new subcommand [trade] which is the home for making
interactions and transactions.

The heart of it lives in [data/ledger.sexp], which acts as a persistent
ledger and memory of every person's accounts and positions. [trade.ml]
contains the new commands, including the [add] and [remove] commands
housed under [admin]. These directly increase and decrease the amount of
money a person has. [buy] and [sell] are the conventional commands
expected to be used more often, which takes in a [token_id] and looks up
its price via the API. It then performs the trade.

An interesting point to note is that if you want to buy a token, you
need to provide [side=sell] to the API, and vice versa. This is because
it hooks you up with a dealer's side. If you wish to buy, you must tell
it to look for a dealer whose [side] is [sell]. This fact is clarified
in the documentation, parameter names, as well as the help command.