Lint Orderbook type
This commit's main goal is to create a [get_book] endpoint. This is so
that we can get the Market from a Token ID.

This would allow us to be able to print the Ledger, which relies on
printing a Portfolio. A portfolio would return Token IDs, which to be
meaningful for a user, needs to be shown as the names of the market.

Hence, a new command has been added in the API and the major changes in
the commit come from [orderbook.ml].