Ensure removal of token if 0 owned
As the commit title suggests, this helps with clutter in the underlying
representation of a [Portfolio]. As a result, the sexp is less
cluttered.

This is as opposed to the [Portfolio] remembering that a user owns 0 of
the token. It deletes it away from the underlying map now.

Tests are added to ensure the above behaviour.