Add get market API query endpoint
This commit finishes the last API call required to re-derive a market
name and token's "Yes" or "No" from only its Token ID.
By querying get-book, and then get-market, we can re-match the token ID
to whether it was a "Yes" or "No".

This commit also miscellaneously does some re-ordering of the variables
in alphabetical order.

There is some reduplication of code, during [Market.of_json], but I deem
it part of design as there is no guarantee that the JSON response has
the same keys. Each file should handle its own JSON.

I considered making them optional arguments and making the polymarket
types default to the sane keys. However, [Polymarket_types] should never
have any knowledge of the API, and the API should solely be handling all
key names.

Perhaps a way to move forward is having a file called
[default_parsers.ml] or something and having reduplication of code
solved that way, but this should only be done if the reduplication gets
even more egregious than it already is.