Add verbose flag to [trade view]
This commit makes the [trade view] command more pleasant to read.
As it currently stands, the command only returns how many of each token
ID a user owns. However, these token IDs are quite meaningless as we do
not know what they are.

Hence, by passing the [-v true] flag now, the command fetches the
relevant APIs and tells us the question and outcome the token ID is
associated with. Furthermore, it also checks for the current price.

This is done by taking the following path using the Polymarket API.
Token ID -> Orderbook's Condition ID -> Market's Question, Tokens Field
-> Search for the Token in the Tokens field.

The previous commits have been made in order to achieve this short-term
goal.

As a result, the [unofficial.ml] file is created to house all
non-official Polymarket API calls. They are convenience functions that
stitch together one or more API calls and have custom logic performed on
them.

As a result, the [to_string] functions of [Portfolio] and [Ledger] now
optionally take in a function that transforms the token IDs to something
meaningful. If [None] is passed, the status quo is retained. If [Some]
function is passed, then a [Deferred.t] is returned.

This also means that the test files must have [Async] opened, which
causes a lot of "dirtying" with [Deferred.t]s strewn around.