Restructure file organisation
This commit is in preparation for adding queries that are not documented
in the official CLOB API. It is related to mirroring the search button
on polymarket. By using the API call to the web itself, we can search
for markets by leveraging their functions.

Because of this, it is likely that something like "polymarket_web" will
need to access [endpoint.ml]. This would signal some kind of wrong file
structuring as [endpoint.ml] should be shared by both the web and the
clob API endpoints.

Hence, this commit creates a central [polymarket/] folder where
everything related to working with the API is housed. Then, in a future
commit, a file or folder can be created which specialises in making web
calls to search for markets.

No code is changed and therefore no functionality is changed. Hence, no
testing is required.