Remove [price] and [winner] field from [Market]
The existence of such fields is not guaranteed by the CLOB API and
should not be included.

This commit is made in preparation of using the web search API to fetch
relevant markets. The returned JSON format for markets differs in this
different API and the provision of [winner] is not explicitly given.
This shines light on the situation and prompts a pre-emptive commit that
removes these fields in preparation.