Add searching of markets query
This commit exposes a new class of API calls. These calls go through the
main Polymarket website, and have no official documentation.

I looked at the Network tab when making requests in the search bar and
used that HTTP request as a base for how [search_markets] works.

Interestingly, the markets returned in a json have different keys, such
as abiding by camelCase instead of snake_case. Furthermore, the
retrieval of the tokens associated with each of the markets is strange.
It sends json formatted in a string, and having each token's data spread
across multiple fields, causing [~tokens] to look quite gnarly in
[web.ml].

It also seems to return bogus placeholder markets with token IDs that
don't exist despite claiming that it is [closed: false]. Hence, extra
filtering must be done to ensure that the market has a non-zero volume.
Then, the only results that remain are those that are actual markets.
We use [is_bogus] for this, using the [active] field, which seems to
separate bogus markets from real ones.

To check the above for yourself, you can try searching for a market with
many options underneath it. For example, the best AI model by x month,
could have many competitors. It will also return "Model A" or "Model B"
being the best AI models as options, which are not buyable.

The new command is housed in [web.ml] and under a [web] subcommand, to
separate it from the official [clob] endpoints.